refactor(cli): split resolution test harness modules
