+++ crates/tsz-cli/src/driver/resolution_tests.rs
@@ -3,6 +3,11 @@ use rustc_hash::{FxHashMap, FxHashSet};
 use tsz::config::{CompilerOptions, resolve_compiler_options};
 use tsz::emitter::ModuleKind;
 
+#[path = "resolution_tests/fixtures.rs"]
+mod fixtures;
+
+use fixtures::*;
+
 #[test]
 fn test_module_resolution_file_exists_cache_is_per_cache() {
     use std::fs;
@@ -225,54 +230,6 @@ fn test_resolve_module_specifier_from_node_modules_package_finds_sibling_package
     );
 
     let _ = fs::remove_dir_all(&dir);
-}
-
-/// Build the shared options used by the pnpm-symlink resolution tests.
-#[cfg(test)]
-fn pnpm_symlink_test_options(preserve_symlinks: bool) -> ResolvedCompilerOptions {
-    ResolvedCompilerOptions {
-        module_resolution: Some(ModuleResolutionKind::Node16),
-        preserve_symlinks,
-        module_suffixes: vec![String::new()],
-        printer: tsz::emitter::PrinterOptions {
-            module: ModuleKind::Node16,
-            ..Default::default()
-        },
-        checker: tsz::checker::context::CheckerOptions {
-            module: ModuleKind::Node16,
-            ..Default::default()
-        },
-        ..Default::default()
-    }
-}
-
-/// Build a pnpm sandbox where the top-level `@types/express` is a symlink into
-/// the `.pnpm` store and `@types/express-serve-static-core` exists *only* as its
-/// transitive (un-hoisted) sibling inside that store. Returns `(project_dir,
-/// sandbox_types_dir)`; `express/index.d.ts` is written with `express_content`.
-#[cfg(test)]
-fn setup_pnpm_express_sandbox(dir_name: &str, express_content: &str) -> (PathBuf, PathBuf) {
-    use std::fs;
-    use std::os::unix::fs::symlink;
-
-    let dir = std::env::temp_dir().join(dir_name);
-    let _ = fs::remove_dir_all(&dir);
-    let sandbox = dir.join("node_modules/.pnpm/@types+express@5.0.6/node_modules/@types");
-    fs::create_dir_all(sandbox.join("express")).unwrap();
-    fs::create_dir_all(sandbox.join("express-serve-static-core")).unwrap();
-    fs::write(sandbox.join("express/index.d.ts"), express_content).unwrap();
-    fs::write(
-        sandbox.join("express-serve-static-core/index.d.ts"),
-        "export type Core = number;",
-    )
-    .unwrap();
-    fs::create_dir_all(dir.join("node_modules/@types")).unwrap();
-    symlink(
-        sandbox.join("express"),
-        dir.join("node_modules/@types/express"),
-    )
-    .unwrap();
-    (dir, sandbox)
 }
 
 /// `/// <reference types="..." />` from inside a pnpm-symlinked `@types/*`
@@ -652,35 +609,6 @@ fn test_duplicate_package_redirects_prefer_stable_lexical_root_when_depth_ties()
     );
 
     let _ = fs::remove_dir_all(&dir);
-}
-
-/// Stand up three duplicate-package copies at the given `node_modules`
-/// sub-paths, each tagged with `name`/`version` so
-/// `build_duplicate_package_redirects` groups them. Returns the per-copy
-/// `index.d.ts` paths in input order.
-fn make_duplicate_package_copies(
-    dir: &Path,
-    name: &str,
-    version: &str,
-    relative_roots: [&str; 3],
-) -> [PathBuf; 3] {
-    let pkg_json = format!(r#"{{"name":"{name}","version":"{version}"}}"#);
-    let mut indices: [PathBuf; 3] = Default::default();
-    for (i, rel) in relative_roots.iter().enumerate() {
-        let pkg_dir = dir.join(rel);
-        std::fs::create_dir_all(&pkg_dir).unwrap();
-        std::fs::write(pkg_dir.join("package.json"), &pkg_json).unwrap();
-        std::fs::write(pkg_dir.join("index.d.ts"), "export {};").unwrap();
-        indices[i] = pkg_dir.join("index.d.ts");
-    }
-    indices
-}
-
-fn dup_pkg_resolver_options() -> ResolvedCompilerOptions {
-    ResolvedCompilerOptions {
-        module_suffixes: vec![String::new()],
-        ..Default::default()
-    }
 }
 
 #[test]
@@ -1103,655 +1031,8 @@ fn test_extensionless_json_import_does_not_resolve_with_resolve_json_module() {
     let _ = fs::remove_dir_all(&dir);
 }
 
-#[test]
-fn test_collect_module_specifiers_finds_dynamic_imports() {
-    let text = r#"import("./foo").then(x => x);"#;
-    let path = Path::new("test.mts");
-    let specifiers = collect_module_specifiers_from_text(path, text);
-    assert!(
-        specifiers.contains(&"./foo".to_string()),
-        "Should find dynamic import specifier './foo', got: {specifiers:?}"
-    );
-}
-
-#[test]
-fn test_collect_module_specifiers_finds_plain_require_calls() {
-    let text = r#"const data = require("./data.json");"#;
-    let path = Path::new("test.js");
-    let specifiers = collect_module_specifiers_from_text(path, text);
-    assert!(
-        specifiers.contains(&"./data.json".to_string()),
-        "Should find require specifier './data.json', got: {specifiers:?}"
-    );
-}
-
-#[test]
-fn test_collect_module_specifiers_finds_require_with_whitespace_before_paren() {
-    let text = r#"const data = require ("./data.json");"#;
-    let path = Path::new("test.js");
-    let specifiers = collect_module_specifiers_from_text(path, text);
-    assert!(
-        specifiers.contains(&"./data.json".to_string()),
-        "Should find spaced require specifier './data.json', got: {specifiers:?}"
-    );
-}
-
-#[test]
-fn test_collect_module_specifiers_finds_jsdoc_import_tags() {
-    let text = r#"
-// @ts-check
-/** @import { "a,b" as CommaName } from "./dep" */
-/** @type {CommaName} */
-const value = "x";
-"#;
-    let path = Path::new("test.js");
-    let specifiers = collect_module_specifiers_from_text(path, text);
-    assert!(
-        specifiers.contains(&"./dep".to_string()),
-        "Should find JSDoc @import specifier './dep', got: {specifiers:?}"
-    );
-}
-
-#[test]
-fn test_collect_jsdoc_import_tag_parses_resolution_mode() {
-    use tsz::module_resolver::ImportingModuleKind;
-    let text = r#"
-// @ts-check
-/** @import { Esm } from "pkg" with { "resolution-mode": "import" } */
-/** @import { Cjs } from "pkg" with { 'resolution-mode': 'require' } */
-/** @import { Plain } from "pkg" */
-/** @type {Esm} */
-const value = "x";
-"#;
-    let path = Path::new("test.js");
-    let requests = collect_module_requests_from_text(path, text);
-
-    // All three target "pkg"; the attribute clause drives the per-request mode.
-    let modes: Vec<Option<ImportingModuleKind>> = requests
-        .iter()
-        .filter(|(specifier, _, _, _)| specifier == "pkg")
-        .map(|(_, _, mode, _)| *mode)
-        .collect();
-
-    assert!(
-        modes.contains(&Some(ImportingModuleKind::Esm)),
-        "expected an ESM-mode request for the `resolution-mode: import` tag, got: {modes:?}"
-    );
-    assert!(
-        modes.contains(&Some(ImportingModuleKind::CommonJs)),
-        "expected a CommonJS-mode request for the `resolution-mode: require` tag, got: {modes:?}"
-    );
-    assert!(
-        modes.contains(&None),
-        "expected a mode-less request for the plain `@import` tag, got: {modes:?}"
-    );
-}
-
-#[test]
-fn test_collect_module_specifiers_require_has_correct_kind() {
-    use tsz::module_resolver::ImportKind;
-    let text = r#"const data = require("./data.json");"#;
-    let file_name = "test.js".to_string();
-    let mut parser = tsz::parser::ParserState::new(file_name, text.to_string());
-    let source_file = parser.parse_source_file();
-    let (arena, _diagnostics) = parser.into_parts();
-    let specifiers = collect_module_specifiers(&arena, source_file);
-    let requires: Vec<_> = specifiers
-        .iter()
-        .filter(|(_, _, kind, _)| *kind == ImportKind::CjsRequire)
-        .map(|(s, _, _, _)| s.as_str())
-        .collect();
-    assert!(
-        requires.contains(&"./data.json"),
-        "Should find CommonJS require, got: {specifiers:?}"
-    );
-}
-
-#[test]
-fn test_collect_module_specifiers_export_import_require_has_correct_kind() {
-    use tsz::module_resolver::ImportKind;
-    let text = r#"export import dep = require("./dep");"#;
-    let file_name = "test.ts".to_string();
-    let mut parser = tsz::parser::ParserState::new(file_name, text.to_string());
-    let source_file = parser.parse_source_file();
-    let (arena, _diagnostics) = parser.into_parts();
-    let specifiers = collect_module_specifiers(&arena, source_file);
-    let requires: Vec<_> = specifiers
-        .iter()
-        .filter(|(_, _, kind, _)| *kind == ImportKind::CjsRequire)
-        .map(|(s, _, _, _)| s.as_str())
-        .collect();
-    assert!(
-        requires.contains(&"./dep"),
-        "Should find exported CommonJS require, got: {specifiers:?}"
-    );
-}
-
-#[test]
-fn test_collect_module_specifiers_dynamic_import_has_correct_kind() {
-    use tsz::module_resolver::ImportKind;
-    let text = r#"import("./foo").then(x => x);"#;
-    let file_name = "test.mts".to_string();
-    let mut parser = tsz::parser::ParserState::new(file_name, text.to_string());
-    let source_file = parser.parse_source_file();
-    let (arena, _diagnostics) = parser.into_parts();
-    let specifiers = collect_module_specifiers(&arena, source_file);
-    let dynamic_imports: Vec<_> = specifiers
-        .iter()
-        .filter(|(_, _, kind, _)| *kind == ImportKind::DynamicImport)
-        .collect();
-    assert_eq!(
-        dynamic_imports.len(),
-        1,
-        "Should find exactly one DynamicImport, got: {specifiers:?}"
-    );
-    assert_eq!(dynamic_imports[0].0, "./foo");
-}
-
-#[test]
-fn test_collect_module_specifiers_import_defer_has_dynamic_import_kind() {
-    use tsz::module_resolver::ImportKind;
-    let text = r#"import.defer("./foo.js").then(x => x);"#;
-    let file_name = "test.mts".to_string();
-    let mut parser = tsz::parser::ParserState::new(file_name, text.to_string());
-    let source_file = parser.parse_source_file();
-    let (arena, _diagnostics) = parser.into_parts();
-    let specifiers = collect_module_specifiers(&arena, source_file);
-    let dynamic_imports: Vec<_> = specifiers
-        .iter()
-        .filter(|(_, _, kind, _)| *kind == ImportKind::DynamicImport)
-        .collect();
-    assert_eq!(
-        dynamic_imports.len(),
-        1,
-        "Should find exactly one import.defer DynamicImport, got: {specifiers:?}"
-    );
-    assert_eq!(dynamic_imports[0].0, "./foo.js");
-}
-
-#[test]
-fn test_module_specifier_detects_type_json_import_attribute() {
-    let text = r#"import data from "./data.json" with { type: "json" };"#;
-    let file_name = "test.mts".to_string();
-    let mut parser = tsz::parser::ParserState::new(file_name, text.to_string());
-    let source_file = parser.parse_source_file();
-    let (arena, _diagnostics) = parser.into_parts();
-    let specifiers = collect_module_specifiers(&arena, source_file);
-    let (specifier, specifier_idx, _, _) = specifiers
-        .iter()
-        .find(|(specifier, _, _, _)| specifier == "./data.json")
-        .expect("expected JSON import specifier");
-
-    assert_eq!(specifier, "./data.json");
-    assert!(
-        module_specifier_has_type_json_import_attribute(&arena, *specifier_idx),
-        "Expected the JSON module specifier to carry a type=json import attribute"
-    );
-}
-
-#[test]
-fn test_collect_module_requests_from_text_carries_type_json_attribute() {
-    let path = Path::new("test.mts");
-    let requests = collect_module_requests_from_text(
-        path,
-        r#"import data from "./data.json" with { type: "json" };"#,
-    );
-    let (_, _, _, has_type_json_attribute) = requests
-        .iter()
-        .find(|(specifier, _, _, _)| specifier == "./data.json")
-        .expect("expected JSON import request");
-
-    assert!(
-        *has_type_json_attribute,
-        "Expected source-discovery module requests to retain type=json import attributes"
-    );
-}
-
-#[test]
-fn test_collect_module_requests_from_text_skips_non_relative_ambient_declaration_names() {
-    let path = Path::new("types.d.ts");
-    let requests = collect_module_requests_from_text(
-        path,
-        r#"
-declare module "*.css" {}
-declare module "virtual:asset" {}
-declare module "./augment" {}
-declare module "pkg" {
-  export { T } from "dep";
-}
-"#,
-    );
-    let specifiers: Vec<_> = requests
-        .iter()
-        .map(|(specifier, _, _, _)| specifier.as_str())
-        .collect();
-
-    assert!(
-        !specifiers.contains(&"*.css"),
-        "ambient wildcard declarations are not source dependencies: {specifiers:?}"
-    );
-    assert!(
-        !specifiers.contains(&"virtual:asset"),
-        "bare ambient declarations are not source dependencies: {specifiers:?}"
-    );
-    assert!(
-        !specifiers.contains(&"pkg"),
-        "ambient declaration names are not source dependencies: {specifiers:?}"
-    );
-    assert!(
-        specifiers.contains(&"./augment"),
-        "relative module augmentation names can target concrete files: {specifiers:?}"
-    );
-    assert!(
-        specifiers.contains(&"dep"),
-        "real re-exports inside ambient module bodies remain dependencies: {specifiers:?}"
-    );
-}
-
-#[test]
-fn test_collect_module_specifiers_for_check_skips_declaration_file_ambient_names() {
-    let text = r#"
-declare module "*.css" {}
-declare module "virtual:asset" {}
-declare module "./augment" {}
-declare module "pkg" {
-  export { T } from "dep";
-}
-"#;
-    let mut parser = tsz::parser::ParserState::new("types.d.ts".to_string(), text.to_string());
-    let source_file = parser.parse_source_file();
-    let (arena, _diagnostics) = parser.into_parts();
-
-    let specifiers: Vec<_> = collect_module_specifiers_for_check(&arena, source_file, true)
-        .into_iter()
-        .map(|(specifier, _, _, _)| specifier)
-        .collect();
-
-    assert!(
-        !specifiers.iter().any(|specifier| specifier == "*.css"),
-        "ambient wildcard declarations are not driver lookups: {specifiers:?}"
-    );
-    assert!(
-        !specifiers
-            .iter()
-            .any(|specifier| specifier == "virtual:asset"),
-        "ambient bare declarations are not driver lookups: {specifiers:?}"
-    );
-    assert!(
-        !specifiers.iter().any(|specifier| specifier == "pkg"),
-        "ambient declaration names are not driver lookups in declaration files: {specifiers:?}"
-    );
-    assert!(
-        specifiers.iter().any(|specifier| specifier == "./augment"),
-        "relative augmentation names still need source-file-specific lookup: {specifiers:?}"
-    );
-    assert!(
-        specifiers.iter().any(|specifier| specifier == "dep"),
-        "real re-exports inside ambient module bodies remain dependencies: {specifiers:?}"
-    );
-}
-
-#[test]
-fn test_collect_module_specifiers_for_check_keeps_bare_source_augmentation_targets() {
-    let text = r#"
-export {};
-declare module "pkg" {
-  export const value: number;
-}
-"#;
-    let mut parser = tsz::parser::ParserState::new("source.ts".to_string(), text.to_string());
-    let source_file = parser.parse_source_file();
-    let (arena, _diagnostics) = parser.into_parts();
-
-    let external_specifiers: Vec<_> =
-        collect_module_specifiers_for_check(&arena, source_file, true)
-            .into_iter()
-            .map(|(specifier, _, _, _)| specifier)
-            .collect();
-    assert!(
-        external_specifiers
-            .iter()
-            .any(|specifier| specifier == "pkg"),
-        "external source augmentations need lookup for TS2664: {external_specifiers:?}"
-    );
-
-    let script_specifiers: Vec<_> = collect_module_specifiers_for_check(&arena, source_file, false)
-        .into_iter()
-        .map(|(specifier, _, _, _)| specifier)
-        .collect();
-    assert!(
-        !script_specifiers.iter().any(|specifier| specifier == "pkg"),
-        "script ambient declarations should not become driver lookups: {script_specifiers:?}"
-    );
-}
-
-#[test]
-fn simple_module_request_scanner_collects_static_imports_and_reexports() {
-    let requests = collect_simple_module_requests_from_text(
-        r#"
-import "./setup";
-import type { Widget } from "./types";
-import view from "./view";
-export { Widget } from "./types";
-export * from "./shared";
-export interface LocalOnly {}
-"#,
-    )
-    .expect("simple static module syntax should not need the source-discovery parser");
-
-    let actual: Vec<_> = requests
-        .iter()
-        .map(|(specifier, kind, _, has_type_json)| (specifier.as_str(), *kind, *has_type_json))
-        .collect();
-    assert_eq!(
-        actual,
-        vec![
-            (
-                "./setup",
-                tsz::module_resolver::ImportKind::EsmImport,
-                false
-            ),
-            (
-                "./types",
-                tsz::module_resolver::ImportKind::EsmImport,
-                false
-            ),
-            ("./view", tsz::module_resolver::ImportKind::EsmImport, false),
-            (
-                "./types",
-                tsz::module_resolver::ImportKind::EsmReExport,
-                false
-            ),
-            (
-                "./shared",
-                tsz::module_resolver::ImportKind::EsmReExport,
-                false
-            ),
-        ]
-    );
-}
-
-#[test]
-fn simple_module_request_scanner_preserves_opposite_quote_specifier_values() {
-    let requests = collect_simple_module_requests_from_text(
-        r#"
-import quotedSingle from "'pkg'";
-export { quotedDouble } from '"pkg"';
-"#,
-    )
-    .expect("opposite quote characters are literal specifier contents");
-
-    let actual: Vec<_> = requests
-        .iter()
-        .map(|(specifier, kind, _, has_type_json)| (specifier.as_str(), *kind, *has_type_json))
-        .collect();
-    assert_eq!(
-        actual,
-        vec![
-            ("'pkg'", tsz::module_resolver::ImportKind::EsmImport, false),
-            (
-                r#""pkg""#,
-                tsz::module_resolver::ImportKind::EsmReExport,
-                false,
-            ),
-        ]
-    );
-}
-
-#[test]
-fn simple_module_request_scanner_falls_back_for_escaped_static_specifiers() {
-    let text = r#"
-import "./d\u0065p";
-export { value } from "./r\u0065exp";
-"#;
-    assert!(
-        collect_simple_module_requests_from_text(text).is_none(),
-        "escaped module specifiers must fall back so the parser can decode them"
-    );
-
-    let requests = collect_module_requests_from_text(std::path::Path::new("index.ts"), text);
-    let actual: Vec<_> = requests
-        .iter()
-        .map(|(specifier, kind, _, _)| (specifier.as_str(), *kind))
-        .collect();
-    assert_eq!(
-        actual,
-        vec![
-            ("./dep", tsz::module_resolver::ImportKind::EsmImport),
-            ("./reexp", tsz::module_resolver::ImportKind::EsmReExport),
-        ]
-    );
-}
-
-#[test]
-fn simple_module_request_scanner_handles_ambient_modules_conservatively() {
-    let simple = collect_simple_module_requests_from_text(
-        r#"
-declare module "*.css" {}
-declare module "virtual:asset" {}
-declare module "./augment" {}
-"#,
-    )
-    .expect("ambient declarations without body imports can stay on the scanner path");
-    let specifiers: Vec<_> = simple
-        .iter()
-        .map(|(specifier, _, _, _)| specifier.as_str())
-        .collect();
-    assert_eq!(specifiers, vec!["./augment"]);
-
-    assert!(
-        collect_simple_module_requests_from_text(
-            r#"
-declare module "pkg" {
-  export { T } from "dep";
-}
-"#
-        )
-        .is_none(),
-        "real dependencies inside ambient module bodies fall back to the parser path"
-    );
-}
-
-#[test]
-fn simple_module_request_scanner_falls_back_for_mode_sensitive_forms() {
-    assert!(
-        collect_simple_module_requests_from_text(
-            r#"import data from "./data.json" with { type: "json" };"#
-        )
-        .is_none()
-    );
-    assert!(
-        collect_simple_module_requests_from_text(r#"const loader = import("./lazy");"#).is_none()
-    );
-    assert!(collect_simple_module_requests_from_text(r#"require("./cjs");"#).is_none());
-}
-
-#[test]
-fn test_collect_module_specifiers_mixed_import_kinds() {
-    use tsz::module_resolver::ImportKind;
-    let text = r#"
-import { foo } from "./static-import";
-import("./dynamic-import");
-export { bar } from "./re-export";
-"#;
-    let file_name = "test.ts".to_string();
-    let mut parser = tsz::parser::ParserState::new(file_name, text.to_string());
-    let source_file = parser.parse_source_file();
-    let (arena, _diagnostics) = parser.into_parts();
-    let specifiers = collect_module_specifiers(&arena, source_file);
-
-    let static_imports: Vec<_> = specifiers
-        .iter()
-        .filter(|(_, _, kind, _)| *kind == ImportKind::EsmImport)
-        .map(|(s, _, _, _)| s.as_str())
-        .collect();
-    assert!(
-        static_imports.contains(&"./static-import"),
-        "Should find static import, got: {static_imports:?}"
-    );
-
-    let dynamic_imports: Vec<_> = specifiers
-        .iter()
-        .filter(|(_, _, kind, _)| *kind == ImportKind::DynamicImport)
-        .map(|(s, _, _, _)| s.as_str())
-        .collect();
-    assert!(
-        dynamic_imports.contains(&"./dynamic-import"),
-        "Should find dynamic import, got: {dynamic_imports:?}"
-    );
-
-    let re_exports: Vec<_> = specifiers
-        .iter()
-        .filter(|(_, _, kind, _)| *kind == ImportKind::EsmReExport)
-        .map(|(s, _, _, _)| s.as_str())
-        .collect();
-    assert!(
-        re_exports.contains(&"./re-export"),
-        "Should find re-export, got: {re_exports:?}"
-    );
-}
-
-#[test]
-fn test_collect_module_specifiers_finds_re_exports_inside_ambient_module_blocks() {
-    use tsz::module_resolver::ImportKind;
-    let text = r#"
-declare module "baz" {
-  export { T } from "foo";
-}
-"#;
-    let file_name = "index.d.ts".to_string();
-    let mut parser = tsz::parser::ParserState::new(file_name, text.to_string());
-    let source_file = parser.parse_source_file();
-    let (arena, _diagnostics) = parser.into_parts();
-    let specifiers = collect_module_specifiers(&arena, source_file);
-
-    let re_exports: Vec<_> = specifiers
-        .iter()
-        .filter(|(_, _, kind, _)| *kind == ImportKind::EsmReExport)
-        .map(|(s, _, _, _)| s.as_str())
-        .collect();
-    assert!(
-        re_exports.contains(&"foo"),
-        "Should find ambient module re-export, got: {specifiers:?}"
-    );
-}
-
-#[test]
-fn test_collect_module_specifiers_extracts_resolution_mode_override() {
-    use tsz::module_resolver::ImportingModuleKind;
-
-    let text = r#"import type { Foo } from "pkg" with { "resolution-mode": "import" };"#;
-    let file_name = "test.ts".to_string();
-    let mut parser = tsz::parser::ParserState::new(file_name, text.to_string());
-    let source_file = parser.parse_source_file();
-    let (arena, _diagnostics) = parser.into_parts();
-    let specifiers = collect_module_specifiers(&arena, source_file);
-
-    assert_eq!(
-        specifiers.len(),
-        1,
-        "Expected exactly one import: {specifiers:?}"
-    );
-    assert_eq!(specifiers[0].0, "pkg");
-    assert_eq!(specifiers[0].3, Some(ImportingModuleKind::Esm));
-}
-
-#[test]
-fn test_collect_module_specifiers_finds_import_type_dependencies() {
-    use tsz::module_resolver::ImportKind;
-
-    let text = r#"export type SomeType = import("./inner").SomeType;"#;
-    let file_name = "index.d.ts".to_string();
-    let mut parser = tsz::parser::ParserState::new(file_name, text.to_string());
-    let source_file = parser.parse_source_file();
-    let (arena, _diagnostics) = parser.into_parts();
-    let specifiers = collect_module_specifiers(&arena, source_file);
-
-    let import_types: Vec<_> = specifiers
-        .iter()
-        .filter(|(_, _, kind, _)| *kind == ImportKind::EsmImport)
-        .map(|(s, _, _, _)| s.as_str())
-        .collect();
-
-    assert!(
-        import_types.contains(&"./inner"),
-        "Should find import type dependency './inner', got: {specifiers:?}"
-    );
-}
-
-#[test]
-fn test_collect_module_specifiers_extracts_import_type_resolution_mode_override() {
-    use tsz::module_resolver::{ImportKind, ImportingModuleKind};
-
-    let text =
-        r#"export type SomeType = import("pkg", { with: { "resolution-mode": "require" } }).Foo;"#;
-    let file_name = "index.ts".to_string();
-    let mut parser = tsz::parser::ParserState::new(file_name, text.to_string());
-    let source_file = parser.parse_source_file();
-    let (arena, _diagnostics) = parser.into_parts();
-    let specifiers = collect_module_specifiers(&arena, source_file);
-
-    let import_types: Vec<_> = specifiers
-        .iter()
-        .filter(|(_, _, kind, _)| *kind == ImportKind::EsmImport)
-        .collect();
-
-    assert_eq!(
-        import_types.len(),
-        1,
-        "Expected one import type, got: {specifiers:?}"
-    );
-    assert_eq!(import_types[0].0, "pkg");
-    assert_eq!(import_types[0].3, Some(ImportingModuleKind::CommonJs));
-}
-
-#[test]
-fn test_collect_module_specifiers_finds_typeof_import_dependencies() {
-    use tsz::module_resolver::ImportKind;
-
-    let text = r#"const parserRef: typeof import("csv-parse") = null as any;"#;
-    let file_name = "index.ts".to_string();
-    let mut parser = tsz::parser::ParserState::new(file_name, text.to_string());
-    let source_file = parser.parse_source_file();
-    let (arena, _diagnostics) = parser.into_parts();
-    let specifiers = collect_module_specifiers(&arena, source_file);
-
-    let import_types: Vec<_> = specifiers
-        .iter()
-        .filter(|(_, _, kind, _)| *kind == ImportKind::EsmImport)
-        .map(|(s, _, _, _)| s.as_str())
-        .collect();
-
-    assert!(
-        import_types.contains(&"csv-parse"),
-        "Should find bare typeof import dependency 'csv-parse', got: {specifiers:?}"
-    );
-}
-
-#[test]
-fn test_collect_module_specifiers_extracts_typeof_import_resolution_mode_override() {
-    use tsz::module_resolver::{ImportKind, ImportingModuleKind};
-
-    let text = r#"type Parser = typeof import("pkg", { with: { "resolution-mode": "require" } });"#;
-    let file_name = "index.ts".to_string();
-    let mut parser = tsz::parser::ParserState::new(file_name, text.to_string());
-    let source_file = parser.parse_source_file();
-    let (arena, _diagnostics) = parser.into_parts();
-    let specifiers = collect_module_specifiers(&arena, source_file);
-
-    let import_types: Vec<_> = specifiers
-        .iter()
-        .filter(|(_, _, kind, _)| *kind == ImportKind::EsmImport)
-        .collect();
-
-    assert_eq!(
-        import_types.len(),
-        1,
-        "Expected one typeof import, got: {specifiers:?}"
-    );
-    assert_eq!(import_types[0].0, "pkg");
-    assert_eq!(import_types[0].3, Some(ImportingModuleKind::CommonJs));
-}
+#[path = "resolution_tests/source_discovery.rs"]
+mod resolution_tests_source_discovery;
 
 #[test]
 fn test_resolve_type_package_entry_with_exports_map() {
